Member management

+++ api/src/controllers/server.controller.ts
@@ -1,6 +1,9 @@
 import { BadRequestException } from '@constants/exceptions/bad-request.exception'
+import { ForbiddenException } from '@constants/exceptions/forbidden.exception'
 import { NotFoundException } from '@constants/exceptions/not-found.exception'
 import { createServerDto, updateServerBannerDto } from '@dtos/server.dto'
+import { MemberRole } from '@prisma/client'
+import MemberService from '@services/member.service'
 import ServerService from '@services/server.service'
 import { NextFunction, Request, Response } from 'express'
 
@@ -9,7 +12,7 @@ import { NextFunction, Request, Response } from 'express'
  */
 export const getServer = async (req: Request, res: Response, next: NextFunction) => {
 
-  const serverId = req.params.id
+  const serverId = req.params.serverId
   if(!serverId) throw new BadRequestException('Server id is required')
 
   const server = await ServerService.findById(serverId)
@@ -31,9 +34,9 @@ export const getUserServers = async (req: Request, res: Response, next: NextFunc
  * @description Create a new server
  */
 export const createServer = async (req: Request, res: Response, next: NextFunction) => {
-  const data = req.body as createServerDto
+  const {name} = req.body as createServerDto
 
-  const server = await ServerService.create(data, req.user!.id)
+  const server = await ServerService.create(name, req.user!.id)
 
   res.status(201).json(server)
 }
@@ -44,11 +47,14 @@ export const createServer = async (req: Request, res: Response, next: NextFuncti
 export const updateServerName = async (req: Request, res: Response, next: NextFunction) => {
   const {name} = req.body as createServerDto
 
-  const serverId = req.params.id
+  const serverId = req.params.serverId
   if(!serverId) throw new BadRequestException('Server id is required')
 
-  const server = await ServerService.findById(req.params.id)
+  const server = await ServerService.findById(serverId)
   if(!server) throw new NotFoundException('Server not found')
+
+  const member = await MemberService.findById(serverId, req.user!.id)
+  if(!member || member.role !== MemberRole.ADMIN) throw new ForbiddenException('You are not allowed to update the server name')
 
   await ServerService.updateName(serverId, name)
 
@@ -63,11 +69,14 @@ export const updateServerName = async (req: Request, res: Response, next: NextFu
 export const updateServerBanner = async (req: Request, res: Response, next: NextFunction) => {
   const {banner} = req.body as updateServerBannerDto
 
-  const serverId = req.params.id
+  const serverId = req.params.serverId
   if(!serverId) throw new BadRequestException('Server id is required')
 
   const server = await ServerService.findById(serverId)
   if(!server) throw new NotFoundException('Server not found')
+
+  const member = await MemberService.findById(serverId, req.user!.id)
+  if(!member || member.role !== MemberRole.ADMIN) throw new ForbiddenException('You are not allowed to update the server banner')
 
   await ServerService.updateBanner(serverId, banner)
 
@@ -80,11 +89,14 @@ export const updateServerBanner = async (req: Request, res: Response, next: Next
  * @description Delete a server
  */
 export const deleteServer = async (req: Request, res: Response, next: NextFunction) => {
-  const serverId = req.params.id
+  const serverId = req.params.serverId
   if(!serverId) throw new BadRequestException('Server id is required')
 
   const server = await ServerService.findById(serverId)
   if(!server) throw new NotFoundException('Server not found')
+
+  const member = await MemberService.findById(serverId, req.user!.id)
+  if(!member || member.role !== MemberRole.ADMIN) throw new ForbiddenException('You are not allowed to delete this server')
 
   await ServerService.delete(serverId)
